requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
PyJWT==2.8.0
Faker==20.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
cryptography==44.0.0
